% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioGeoBias_call_map_api.R
\name{call_map_api}
\alias{call_map_api}
\title{Call map API for bias grid}
\usage{
call_map_api(taxonkey = 1, decades = c("1990_2000", "2000_2010",
  "2010_2020"), layers = c("OBS", "SP", "OTH"), living = TRUE,
  fossil = FALSE, breaks = NULL, nbreaks = NULL, x = 0, y = 0,
  z = 0)
}
\arguments{
\item{taxonkey}{a valid GBIF taxon key as returned by ###}

\item{decades}{optional parameter to restrict the results to a certain time
period, specified as decades in the format: c("1990_2000","2000_2010"). If
no value is given, the value of c("1990_2000","2000_2010","2010_2020"), i.e.
the time from 1990 to present day is used by default.}

\item{layers}{optional parameter to specify the types of observations to
consider, specified as any combination of "OBS" (observations), "SP"
(specimen), and "OTH" (other). If no value is given, the value of
c("OBS","SP","OTH") is used by default.}

\item{living}{optional logical parameter to specify if living specimen should
be included in the results. Default is TRUE.}

\item{fossil}{optional logical parameter to specify if fossils should be
included in the results. Default is FALSE.}

\item{breaks}{optional numeric vector to specify the break points of value
bins that shouls be used to represent species occurrence record numbers in
the resulting raster. This should be 50 or lower, but span the spectrum of
expected number of observations per raster cell to provide maximum
contrast between observed values. If you are unsure about what to specify,
leave this empty and get results with the standard settings first. In this
case, BioGeoBias will try to generate a suitable exponential sequence of
break values estimated from the total number of species occurrence records
available for yours species.}

\item{nbreaks}{optional numeric parameter to specify the number of breaks to
use for the binning of raster values. Should be between 10 and 50. If breaks
is specified this values will be ignored. Default is 50.}

\item{x}{numeric value to specify the longitudinal position of the raster
extent. Default is 0.}

\item{y}{numeric value to specify the latitudinal position of the raster
extent. Default is 0.}

\item{z}{numeric value to specify the zoom factor of the raster extent.
Default is 0 for global extent.}

\item{latMax}{maximum latitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}
}
\description{
This function takes a valid GBIF taxon key to identify a
target group and returns a raster representing number of species occurrence
records per raster cell for the target group. This raster can be used as
bias grid for species distribution models of a species in that group.
}
\details{
This function uses the arguments passed on to generate a query
to the GBIF web map API. The API returns a web tile object that can be read
as a png and converted into an R raster object. The break values or nbreaks
generate a custom colour palette for the web tile, with each bin
corresponding to one grey value. After retrieval, the raster is reclassified
to the actual break values. This is a somewhat hacky but nonetheless
functional solution in the absence of a GBIF raster API or an rgbif map API
implementation.
}
\author{
Jan Laurens Geffert, \email{laurensgeffert@gmail.com}
}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\url{http://www.ecography.org/accepted-article/performance-tradeoffs-target-group-bias-correction-species-distribution-models}
}
\keyword{GBIF}
\keyword{bias,}
\keyword{grid,}
\keyword{map}
\keyword{raster,}
\keyword{sampling}
\keyword{tile,}
\keyword{web}

