% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioGeoBias_create_bias_grid.R
\name{create_bias_grid}
\alias{create_bias_grid}
\title{Create a Bias Grid from GBIF web map API}
\usage{
create_bias_grid(taxonkey = 1, target_rank = rgbif::taxrank(),
  kingdom = c(NULL, "animalia", "plantae", "archaea", "bacteria", "fungi",
  "protozoa", "viruses"), lonMin = NULL, lonMax = NULL, latMin = NULL,
  latMax = NULL, breaks = NULL, nbreaks = NULL)
}
\arguments{
\item{target_rank}{(character) a taxon rank for the target group (optional).}

\item{kingdom}{(character) the kingdom of the species. Can be supplied to
avoid possible confusion when matching names. Should be one of c('animalia',
'plantae','archaea','bacteria','fungi','protozoa','viruses') (optional).}

\item{lonMin}{minimum longitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}

\item{lonMax}{maximum longitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}

\item{latMin}{minimum latitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}

\item{latMax}{maximum latitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}

\item{species_name}{(character) a valid taxon name, presumably a species name.}
}
\description{
This function takes a species name and a rank, and returns a
raster of number of species occurrence records per cell which can be used as
a sampling grid for bias correction in SDM applications.
}
\details{
hello
}
\author{
Jan Laurens Geffert, \email{laurensgeffert@gmail.com}
}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\url{http://www.ecography.org/accepted-article/performance-tradeoffs-target-group-bias-correction-species-distribution-models}
}
\keyword{GBIF,}
\keyword{bias,}
\keyword{grid}
\keyword{map}
\keyword{raster,}
\keyword{sampling}
\keyword{tile,}
\keyword{web}

