% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioGeoBias_create_target_group_background_data.R
\name{create_target_group_background_data}
\alias{create_target_group_background_data}
\title{Get background data using the Target Group Background Approach (TGBA)}
\usage{
create_target_group_background_data(species_name,
  target_rank = rgbif::taxrank(), kingdom = c(NULL, "animalia", "plantae",
  "archaea", "bacteria", "fungi", "protozoa", "viruses"), lonMin = NULL,
  lonMax = NULL, latMin = NULL, latMax = NULL, yearMin = NULL,
  yearMax = NULL, limit = 2e+05)
}
\arguments{
\item{species_name}{(character) a valid taxon name, presumably a species name.}

\item{target_rank}{(character) a taxon rank for the target group (optional).}

\item{kingdom}{(character) the kingdom of the species. Can be supplied to
avoid possible confusion when matching names. Should be one of c('animalia',
'plantae','archaea','bacteria','fungi','protozoa','viruses') (optional).}

\item{lonMin}{minimum longitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}

\item{lonMax}{maximum longitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}

\item{latMin}{minimum latitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}

\item{latMax}{maximum latitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}
}
\description{
This function takes a species name and a rank, and returns a set
of background data for species distribution models of the species using the
Target Group Background Approach (TGBA).
}
\details{
With the TGBA, occurrence records of other species in the given
taxonomic group are considered background points. Assuming that sampling
intensity is spatially evenly distributed across all species in the group
these occurrences become a proxy of the spatial distribution of sampling
effort. Using them as background data results in models that explicitly
account for sampling bias and can create improved predictions of species
distributions.
The function connects to the GBIF API using functions from the R
package rgbif. First, the name of the species is matched against the GBIF
backbone taxonomy. Next, the name key of the taxon corresponding to the
parent taxon of the given rank is extracted. An occurrence search for
georeferenced records of that taxon returns the dataset we want.
}
\author{
Jan Laurens Geffert, \email{laurensgeffert@gmail.com}
}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\url{http://www.ecography.org/accepted-article/performance-tradeoffs-target-group-bias-correction-species-distribution-models}
}
\keyword{SDM}
\keyword{TGBA,}
\keyword{bias,}
\keyword{sampling}

