% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioGeoBias_bias_correction.R
\name{bias_correction}
\alias{bias_correction}
\title{Bias Correction}
\usage{
bias_correction(species_name, target_rank = rgbif::taxrank(),
  kingdom = c(NULL, "animalia", "plantae", "archaea", "bacteria", "fungi",
  "protozoa", "viruses"), lonMin = NULL, lonMax = NULL, latMin = NULL,
  latMax = NULL, breaks = NULL, nbreaks = NULL, limit = 2e+05)
}
\arguments{
\item{species_name}{(character) a valid taxon name, presumably a species name.}

\item{target_rank}{(character) a taxon rank for the target group (optional).}

\item{kingdom}{(character) the kingdom of the species. Can be supplied to
avoid possible confusion when matching names. Should be one of c('animalia',
'plantae','archaea','bacteria','fungi','protozoa','viruses') (optional).}

\item{lonMin}{minimum longitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}

\item{lonMax}{maximum longitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}

\item{latMin}{minimum latitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}

\item{latMax}{maximum latitude of a rectancular bounding box restricting the
search for species occurrences of the target group (optional).}
}
\description{
Choose bias correction approach and generate bias correction data
}
\details{
This function takes a species name and a rank, and returns a
dataset suitable for bias correction by invoking
\code{create_target_group_background_data} or
\code{create_bias_grid}, depending on the number of species occurrence
records available for the species and the spatial extent of the selection.
}
\author{
Jan Laurens Geffert, \email{laurensgeffert@gmail.com}
}
\keyword{GBIF,}
\keyword{SDM}
\keyword{bias,}
\keyword{sampling}

